==> 01-data-creation.ipynb <==
## Notebook to Download raw data to the Working Folder

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import glob
import azureml.core

from azureml.core import Environment, Experiment, Workspace
from azureml.core.compute import ComputeTarget, AmlCompute
from azureml.core.runconfig import RunConfiguration
from azureml.core.conda_dependencies import CondaDependencies
from azureml.pipeline.steps import PythonScriptStep
from azureml.pipeline.core import Pipeline, PublishedPipeline, PipelineEndpoint
from azureml.pipeline.core.schedule import ScheduleRecurrence, Schedule, TimeZone
from azureml.core import Run, Datastore, Dataset, Workspace

from azureml.widgets import RunDetails

from IPython.core.display import display, HTML

# display(HTML("<style>.container { width:100% !important; }</style>"))


ws = Workspace.from_config()
ds = ws.get_default_datastore()
# ds_weblogs = Datastore.get(ws, 'weblogs')

# cpu_cluster_name = 'abir-forecasting'

print('Workspace Name: ' + ws.name, 
      'Azure Region: ' + ws.location, 
      'Subscription Id: ' + ws.subscription_id, 
      'Resource Group: ' + ws.resource_group, 
      sep = '\n')

# history = Datastore.get(ws, "gmr-history-data")
datastores = ws.datastores
for name, datastore in datastores.items():
    print(name, datastore.datastore_type)

blob_datastore_name='gmr' # Name of the datastore to workspace
container_name=os.getenv("BLOB_CONTAINER", "gmr-history-data") # Name of Azure blob container
account_name=os.getenv("BLOB_ACCOUNTNAME", "ptplstorage") # Storage account name
account_key=os.getenv("BLOB_ACCOUNT_KEY", "+rs0KeAAF8+h/dwTB6rfbc1s11yxnrNLIytt3TWucvxdHnMIqiDQAodHzVyacOwNoXBAnIrpYkuUEvXA2elztw==") # Storage account access key

blob_datastore = Datastore.register_azure_blob_container(workspace=ws, 
                                                         datastore_name=blob_datastore_name, 
                                                         container_name=container_name, 
                                                         account_name=account_name,
                                                         account_key=account_key)

# Get a named datastore from the current workspace
datastore = Datastore.get(ws, datastore_name='gmr')

datastore

dataset = Dataset.File.from_files(path = [(datastore, 'GMR_Alarms/*.xlsx')])
dataset.download('/mnt/batch/tasks/shared/LS_root/mounts/clusters/abir-msft/code/Abir/data/Alarms')

dataset = Dataset.File.from_files(path = [(datastore, 'Inverter/*.csv')])
dataset.download('/mnt/batch/tasks/shared/LS_root/mounts/clusters/abir-msft/code/Abir/data/Inverter')

dataset = Dataset.File.from_files(path = [(datastore, 'SMB/*.csv')])
dataset.download('/mnt/batch/tasks/shared/LS_root/mounts/clusters/abir-msft/code/Abir/data/SMB')

! ls /tmp/tmppugh863e
